chore: renames pkg/apis to pkg/api
