+++ pkg/api/addtoscheme_maistra_v1alpha1.go
@@ -1,7 +1,7 @@
-package apis
+package api
 
 import (
-	"github.com/maistra/istio-workspace/pkg/apis/maistra/v1alpha1"
+	"github.com/maistra/istio-workspace/pkg/api/maistra/v1alpha1"
 )
 
 func init() {
