+++ pkg/client/clientset/versioned/fake/register.go
@@ -3,7 +3,7 @@
 package fake
 
 import (
-	maistrav1alpha1 "github.com/maistra/istio-workspace/pkg/apis/maistra/v1alpha1"
+	maistrav1alpha1 "github.com/maistra/istio-workspace/pkg/api/maistra/v1alpha1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	runtime "k8s.io/apimachinery/pkg/runtime"
 	schema "k8s.io/apimachinery/pkg/runtime/schema"
