+++ pkg/client/clientset/versioned/typed/maistra/v1alpha1/maistra_client.go
@@ -3,7 +3,7 @@
 package v1alpha1
 
 import (
-	v1alpha1 "github.com/maistra/istio-workspace/pkg/apis/maistra/v1alpha1"
+	v1alpha1 "github.com/maistra/istio-workspace/pkg/api/maistra/v1alpha1"
 	"github.com/maistra/istio-workspace/pkg/client/clientset/versioned/scheme"
 	rest "k8s.io/client-go/rest"
 )
