+++ pkg/client/clientset/versioned/typed/maistra/v1alpha1/session.go
@@ -6,7 +6,7 @@ import (
 	"context"
 	"time"
 
-	v1alpha1 "github.com/maistra/istio-workspace/pkg/apis/maistra/v1alpha1"
+	v1alpha1 "github.com/maistra/istio-workspace/pkg/api/maistra/v1alpha1"
 	scheme "github.com/maistra/istio-workspace/pkg/client/clientset/versioned/scheme"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	types "k8s.io/apimachinery/pkg/types"
