+++ pkg/controller/session/model_convert.go
@@ -3,7 +3,7 @@ package session
 import (
 	"reflect"
 
-	istiov1alpha1 "github.com/maistra/istio-workspace/pkg/apis/maistra/v1alpha1"
+	istiov1alpha1 "github.com/maistra/istio-workspace/pkg/api/maistra/v1alpha1"
 	"github.com/maistra/istio-workspace/pkg/model"
 )
