+++ pkg/controller/session/session_controller_int_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/maistra/istio-workspace/pkg/template"
 
-	"github.com/maistra/istio-workspace/pkg/apis/maistra/v1alpha1"
+	"github.com/maistra/istio-workspace/pkg/api/maistra/v1alpha1"
 	"github.com/maistra/istio-workspace/pkg/controller/session"
 	"github.com/maistra/istio-workspace/pkg/log"
 	"github.com/maistra/istio-workspace/test"
