+++ pkg/internal/session/session_client.go
@@ -3,7 +3,7 @@ package session
 import (
 	"context"
 
-	istiov1alpha1 "github.com/maistra/istio-workspace/pkg/apis/maistra/v1alpha1"
+	istiov1alpha1 "github.com/maistra/istio-workspace/pkg/api/maistra/v1alpha1"
 	"github.com/maistra/istio-workspace/pkg/client/clientset/versioned"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
