+++ pkg/internal/session/session_client_test.go
@@ -1,7 +1,7 @@
 package session_test
 
 import (
-	istiov1alpha1 "github.com/maistra/istio-workspace/pkg/apis/maistra/v1alpha1"
+	istiov1alpha1 "github.com/maistra/istio-workspace/pkg/api/maistra/v1alpha1"
 	testclient "github.com/maistra/istio-workspace/pkg/client/clientset/versioned/fake"
 	"github.com/maistra/istio-workspace/pkg/internal/session"
