+++ pkg/internal/session/session_test.go
@@ -3,7 +3,7 @@ package session_test
 import (
 	"time"
 
-	istiov1alpha1 "github.com/maistra/istio-workspace/pkg/apis/maistra/v1alpha1"
+	istiov1alpha1 "github.com/maistra/istio-workspace/pkg/api/maistra/v1alpha1"
 	testclient "github.com/maistra/istio-workspace/pkg/client/clientset/versioned/fake"
 	"github.com/maistra/istio-workspace/pkg/internal/session"
